Properly mock the firestore module
