+++ src/__mocks__/firebase.ts
@@ -1,0 +1,33 @@
+import { designers } from "../testData/designers";
+import tags from "../testData/tags";
+
+const collection = (name: string) => {
+  const collections: { [index: string]: any } = {
+    tags: {
+      get: () => {
+        return new Promise(resolve => {
+          resolve(tags);
+        });
+      }
+    },
+    people: {
+      get: () => {
+        return new Promise(resolve => {
+          resolve(designers);
+        });
+      }
+    }
+  };
+
+  return collections[name];
+};
+
+export const firestore = {
+  collection
+};
+
+const app = {
+  firestore: () => firestore
+}
+
+export default app;
