--- src/__mocks__/firestore.ts
@@ -1,19 +1,0 @@
-const collection = (name: string) => {
-  const collections: { [index: string]: any } = {
-    people: {
-      get: () => {
-        return new Promise(resolve => {
-          resolve([]);
-        });
-      }
-    }
-  };
-
-  return collections[name];
-};
-
-const firestore = {
-  collection
-};
-
-export default firestore;
